Update Sarana Fix

+++ app/Http/Controllers/SaranasController.php
@@ -18,11 +18,11 @@ class SaranasController extends Controller
     public function index($id)
     {
         $data_perumahan = Perumahans::find($id);
-        $data_sarana = Sarana::where('perumahan_id',$id)->get();
-        $data_foto_sarana = FotoSarana::where('perumahan_id',$id)->get();
-        $data_koordinat_sarana = KoordinatSarana::where('perumahan_id',$id)->get();
-        return view('PSU_Perumahan.sarana.index', compact('data_perumahan','data_sarana',
-        'data_foto_sarana','data_koordinat_sarana'));
+        $data_sarana = Sarana::where('perumahan_id', $id)->get();
+        $data_foto_sarana = FotoSarana::where('perumahan_id', $id)->get();
+        $data_koordinat_sarana = KoordinatSarana::where('perumahan_id', $id)->get();
+        return view('PSU_Perumahan.sarana.index', compact('data_perumahan', 'data_sarana',
+            'data_foto_sarana', 'data_koordinat_sarana'));
     }
 
     /**
@@ -38,7 +38,7 @@ class SaranasController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param \Illuminate\Http\Request $request
      * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
      */
     public function store(Request $request)
@@ -59,20 +59,20 @@ class SaranasController extends Controller
 
         $this->validate($request, $rules, $customMessages);
 
-        foreach ($request->data_sarana as $key => $value){
+        foreach ($request->data_sarana as $key => $value) {
             Sarana::create($value);
         }
 
         $perumahan_id = $request->data_sarana[0]['perumahan_id'];
         return redirect()->action('PerumahansController@edit', ['id' => $perumahan_id])
-            ->with('status','Data Sarana Berhasil Disimpan');
+            ->with('status', 'Data Sarana Berhasil Disimpan');
 
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  \App\Sarana  $sarana
+     * @param \App\Sarana $sarana
      * @return \Illuminate\Http\Response
      */
     public function show(Sarana $sarana)
@@ -83,7 +83,7 @@ class SaranasController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Sarana  $sarana
+     * @param \App\Sarana $sarana
      * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\View\View
      */
     public function edit(Sarana $sarana)
@@ -94,8 +94,8 @@ class SaranasController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Sarana  $sarana
+     * @param \Illuminate\Http\Request $request
+     * @param \App\Sarana $sarana
      * @return \Illuminate\Http\RedirectResponse
      */
     public function update(Request $request, Sarana $sarana)
@@ -118,23 +118,30 @@ class SaranasController extends Controller
             'luas_sarana' => $request->luas_sarana,
             'kondisi_sarana' => $request->kondisi_sarana
         ]);
-        return redirect()->action('SaranasController@index', ['id' => $perumahan_id])
-            ->with('status','Data Dengan ID '.$sarana->id.' Berhasil Di Update');
+        return redirect()->action('PerumahansController@edit', ['id' => $perumahan_id])
+            ->with('status', 'Data Sarana Dengan ID' . $sarana->id . ' Berhasil Di Update');
 
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Sarana  $sarana
+     * @param \App\Sarana $sarana
      * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
      */
     public function destroy(Request $request, Sarana $sarana)
     {
-
         $perumahan_id = $request->get('perumahan_id');
-        Sarana::destroy($sarana->id);
+        $data_file = FotoSarana::where('sarana_id', $sarana->id)->get();
+        foreach ($data_file as $file) {
+            $path = public_path('/assets/uploads/perumahan/sarana/'). $file->file_foto;
+            if (file_exists($path)) {
+                unlink($path);
+                FotoSarana::where('sarana_id', $sarana->id)->delete();
+                Sarana::destroy($sarana->id);
+            }
+        }
         return redirect()->action('PerumahansController@edit', ['id' => $perumahan_id])
-            ->with('status','Data Sarana Dengan ID '.$sarana->id.' Berhasil Dihapus');
+            ->with('status', 'Data Sarana Dengan ID ' . $sarana->id . ' Berhasil Dihapus');
     }
 }
